chore: 엔티티 임시 패키지(draftDomain)으로이동
