+++ src/main/java/com/odorok/OdorokApplication/draftDomain/ArticleImage.java
@@ -1,4 +1,4 @@
-package com.odorok.OdorokApplication.infrastructures.domain;
+package com.odorok.OdorokApplication.draftDomain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -6,16 +6,19 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Entity(name = "diseases")
+@Entity(name = "article_images")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class Disease {
+public class ArticleImage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "name")
-    private String name;
+    @Column(name = "img_url")
+    private String imgUrl;
+
+    @Column(name = "article_id")
+    private Long articleId;
 }
