+++ src/main/java/com/odorok/OdorokApplication/draftDomain/Disease.java
@@ -1,24 +1,21 @@
-package com.odorok.OdorokApplication.infrastructures.domain;
+package com.odorok.OdorokApplication.draftDomain;
+
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Entity(name = "profile_deco")
+@Entity(name = "diseases")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class ProfileDeco {
-
+public class Disease {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "profile_id")
-    private Long profileId;
-
-    @Column(name = "item_id")
-    private Long itemId;
+    @Column(name = "name")
+    private String name;
 }
