+++ src/main/java/com/odorok/OdorokApplication/draftDomain/ProfileDeco.java
@@ -1,21 +1,24 @@
-package com.odorok.OdorokApplication.infrastructures.domain;
-
+package com.odorok.OdorokApplication.draftDomain;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Entity(name = "contenttypes")
+@Entity(name = "profile_deco")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class ContentType {
+public class ProfileDeco {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "name")
-    private String name;
+    @Column(name = "profile_id")
+    private Long profileId;
+
+    @Column(name = "item_id")
+    private Long itemId;
 }
